Preparativos para la creacion del dashboard (En proceso)
Se arreglo la redireccion del login, ahora lleva a la ruta /administracion
Se creo un trait para los botones de edit y destroy para los modelos
Se creo los componentes editar y delete. Aun en proceso (VueJs)

PRODUCTOS
Se arreglo un error en la migracion con las relaciones.
Se integro el nuevo trait y se hizo los test correspondientes.
Se creo la página del index

Maqueta
Partials de Breadcrumbs y tabla del index.
Arreglo en el sidebar que se veia en el login

+++ database/migrations/2019_10_27_134004_create_products_table.php
@@ -21,10 +21,10 @@ class CreateProductsTable extends Migration
             $table->float('price', 8, 2);
             $table->boolean('published')->default(0);
 
-            $table->unsignedInteger('category_id')->nullable();
+            $table->unsignedBigInteger('category_id')->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
 
-            $table->unsignedInteger('brand_id')->nullable();
+            $table->unsignedBigInteger('brand_id')->nullable();
             $table->foreign('brand_id')->references('id')->on('brands');
             $table->timestamps();
         });
